Major fix to font resources not being applied.
Lots of Android and minor Linux font/position fixes.
Remove menu button in favor of left swipe gesture.
Replace traffic range zoom in/out with swipe up/down gesture.
Extend altitude box and make it semi-transparent for low range vertical speed movements.

+++ FuelTanksDialog.cpp
@@ -130,7 +130,7 @@ void FuelTanksDialog::resetFuel()
     keypadL.setGeometry( 0, 0, 400, 350 );
     keypadL.exec();
 
-    iGalL = keypadL.value();
+    iGalL = static_cast<int>( keypadL.value() );
 
     if( m_tanks.bDualTanks )
     {
@@ -139,7 +139,7 @@ void FuelTanksDialog::resetFuel()
         keypadR.setGeometry( 0, 0, 400, 350 );
         keypadR.exec();
 
-        iGalR = keypadR.value();
+        iGalR = static_cast<int>( keypadR.value() );
         if( iGalR < 0 )
             iGalR = 0;
     }
